split admins and templates
