+++ opencivicdata/core/admin/__init__.py
@@ -1,0 +1,1 @@
+from . import other, organization, person   # noqa
