+++ opencivicdata/core/templates/opencividata/admin/merge.html
