+++ opencivicdata/legislative/admin/__init__.py
@@ -1,0 +1,1 @@
+from . import bill, vote, event     # noqa
