+++ opencivicdata/legislative/admin/vote.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
+from opencivicdata.core.admin.base import ModelAdmin, ReadOnlyTabularInline
 from .. import models
-from .base import ModelAdmin, ReadOnlyTabularInline
 
 
 class VoteCountInline(ReadOnlyTabularInline):
